backend/app/requirements.txt
fastapi
uvicorn[standard]
elasticsearch==8.13.2
python-dotenv
pydantic
